Fix commit message formatting
